Attempt magicsequence
TLE

+++ magicsequence.rs
@@ -1,17 +1,17 @@
 use std::io::{self, BufRead};
 
 struct MagicSequence {
-        curr: u32,
+        curr: u64,
         i: usize,
-        a: Vec<u32>,
-        b: u32,
-        c: u32,
+        a: Vec<u64>,
+        b: u64,
+        c: u64,
 }
 
 impl Iterator for MagicSequence {
-    type Item = u32;
+    type Item = u64;
 
-    fn next(&mut self) -> Option<u32> {
+    fn next(&mut self) -> Option<u64> {
         if self.i < self.a.len() {
             let old = self.curr;
             self.curr = (old * self.b + self.a[self.i]) % self.c;
@@ -31,7 +31,7 @@ impl Iterator for MagicSequence {
     }
 }
 
-fn magicsequence(a: Vec<u32>, b: u32, c:u32, hash: bool) -> MagicSequence {
+fn magicsequence(a: Vec<u64>, b: u64, c:u64, hash: bool) -> MagicSequence {
     MagicSequence {
         curr: if hash { a[0] % c } else { a[0] },
         i: 1,
@@ -45,25 +45,25 @@ fn main() {
 
     let mut lines = stdin.lock().lines();
 
-    let tc = lines.next().unwrap().unwrap().parse::<u32>().unwrap();
+    let tc = lines.next().unwrap().unwrap().parse::<u64>().unwrap();
 
     for _ in 0..tc {
         let n = lines.next().unwrap().unwrap().parse::<usize>().unwrap();
         let abc = lines.next().unwrap().unwrap().split(" ")
-            .map(|x| x.parse::<u32>().unwrap()).collect::<Vec<_>>();
+            .map(|x| x.parse::<u64>().unwrap()).collect::<Vec<_>>();
         let (a, b, c) = (abc[0], abc[1], abc[2]);
         let xy = lines.next().unwrap().unwrap().split(" ")
-            .map(|x| x.parse::<u32>().unwrap()).collect::<Vec<_>>();
+            .map(|x| x.parse::<u64>().unwrap()).collect::<Vec<_>>();
         let (x, y) = (xy[0], xy[1]);
 
         // Compute S
-        let mut s = magicsequence(vec![a; n], b, c, false).collect::<Vec<u32>>();
+        let mut s = magicsequence(vec![a; n], b, c, false).collect::<Vec<u64>>();
 
         // Sort S
         s.sort();
 
         // Compute hash
-        let hash: u32 = magicsequence(s, x, y, true).last().unwrap();
+        let hash: u64 = magicsequence(s, x, y, true).last().unwrap();
 
         println!("{}", hash);
     }
